feat: connect wallet
